@@ -12,18 +12,18 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import java.util.UUID;
 
 @RestController
 @RequestMapping("/companies")
 public class CompanyController {
     private final ICompanyService companyService;
-    private final IFileService fileService;
+//    private final IFileService fileService;
 
-    public CompanyController(ICompanyService companyService, IFileService fileService) {
+    public CompanyController(ICompanyService companyService/*, IFileService fileService*/) {
         this.companyService = companyService;
-        this.fileService = fileService;
+//        this.fileService = fileService;
     }
 
     @PostMapping
@@ -42,14 +42,14 @@ public class CompanyController {
         return new ResponseEntity<>(companyService.findById(id), HttpStatus.OK);
     }
 
-    @PostMapping(
-            path = "/file",
-            consumes = MediaType.MULTIPART_FORM_DATA_VALUE,
-            produces = MediaType.APPLICATION_JSON_VALUE
-    )
-    public ResponseEntity<File> saveTodo(@RequestParam("title") String title,
-                                         @RequestParam("description") String description,
-                                         @RequestParam("file") MultipartFile file) {
-        return new ResponseEntity<>(fileService.saveFile(title, description, file), HttpStatus.OK);
-    }
+//    @PostMapping(
+//            path = "/file",
+//            consumes = MediaType.MULTIPART_FORM_DATA_VALUE,
+//            produces = MediaType.APPLICATION_JSON_VALUE
+//    )
+//    public ResponseEntity<File> saveTodo(@RequestParam("title") String title,
+//                                         @RequestParam("description") String description,
+//                                         @RequestParam("file") MultipartFile file) {
+//        return new ResponseEntity<>(fileService.saveFile(title, description, file), HttpStatus.OK);
+//    }
 }
